scenarios/chatbot/normal-chatbot/pyproject.toml
[project]
name = "normal-chatbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain>=1.0.3",
    "langchain-community>=0.4.1",
    "langchain-core>=1.0.2",
    "langchain-openai>=1.0.1",
    "langgraph>=1.0.2",
    "python-dotenv>=1.2.1",
]
